refactor : step5 페이지 리팩토링
