+++ app/join/funnels/step4/components/ImageForm.client.tsx
@@ -5,7 +5,6 @@ import ImageFrame from '@/components/common/ImageFrame';
 export default function ImageForm() {
   const { setValue, watch } = useJoinContext();
   const setImageUrl = (imageUrl: string) => setValue('imageUrl', imageUrl);
-  console.log(watch('imageUrl'));
 
   return <ImageFrame setImageUrl={setImageUrl} />;
 }
