+++ app/join/funnels/step5/Step5Component.server.tsx
@@ -1,4 +1,5 @@
 import FormSection from './components/FormSection.client';
+import SubmitSection from './components/SubmitSection.client';
 import JoinContentTextMessage from '../../components/JoinContentTextMessage.server';
 import JoinTitleTextMessage from '../../components/JoinTitleTextMessage.server';
 import { Spacing } from '@/components/common/Spacing';
@@ -14,6 +15,7 @@ export default function Step5Component() {
       <JoinContentTextMessage>3개를 선택해주세요.</JoinContentTextMessage>
       <Spacing size={30} />
       <FormSection />
+      <SubmitSection />
     </main>
   );
 }
